Split sisr/utils.py into sisr/colorspace.py and sisr/cache.py
utils.py held two unrelated responsibilities: BT.601 colorspace math (needs
torch) and the LMDB cache (needs only lmdb + tqdm). Because they shared a
module, `import sisr.utils` cost 3.78s instead of 0.31s.

That is paid per spawned process, not once. parallel_build fans out over a
ProcessPoolExecutor and Windows/macOS re-import the worker's module tree on
spawn, so every build worker imported a torch it never calls. A real cache
build resolves num_workers to min(os.cpu_count(), n_items) — 16 on this box —
burning ~60 CPU-seconds of import per build.

Clean break, no compatibility shim: sisr.utils is gone and every importer
moved. tests/test_utils.py splits along the same seam. test_cache.py gains a
guard asserting sisr.cache imports no torch, so the coupling cannot silently
return.

Measured: tests/test_cache.py::test_parallel_build_persists_every_submitted_item
4.63s -> 1.11s. 217 tests pass.
