+++ sisr/datasets/srcnn.py
@@ -3,7 +3,7 @@
 LR is generated from HR via blur → bicubic-down → bicubic-up so the LR
 patches share the spatial size of the HR patches (pre-upsampled SRCNN
 formulation). :class:`TrainDataset` caches sliding-window sub-images
-through :class:`~sisr.utils.LMDBCache`; :class:`ValidationDataset`
+through :class:`~sisr.cache.LMDBCache`; :class:`ValidationDataset`
 generates LR pairs on the fly for full images.
 """
 
@@ -18,7 +18,7 @@ import numpy as np
 import torch
 from PIL import Image
 
-from ..utils import LMDBCache, LMDBCacheBuildContext
+from ..cache import LMDBCache, LMDBCacheBuildContext
 from .base import SRDataset
