+++ sisr/processors/ycbcr.py
@@ -2,7 +2,7 @@
 
 import torch
 
-from sisr.utils import rgb_to_ycbcr, ycbcr_to_rgb
+from sisr.colorspace import rgb_to_ycbcr, ycbcr_to_rgb
 
 from .base import SRProcessor
